[XLA:GPU] Generalize and unify the autotuning map
Remove duplication in code handling the serialization of the autotuning DB.

The logic for autotuning of cuBLAS gemms is somewhat complicated and redundant,
but due to it going away in Ampere+, this CL tries to minimize the number of
changes.

Makes previous autotuning dumps invalid, as the AOT format changes.

PiperOrigin-RevId: 539107671